test(providers): remove the shadow lifecycle provider
